requirements.txt
alabaster==0.7.10
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0
django-widgets-improved==1.5.0
djangorestframework==3.7.7
docutils==0.14
gunicorn==19.7.1
idna==2.6
imagesize==0.7.1
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
Pygments==2.2.0
PyJWT==1.5.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
snowballstemmer==1.2.1
social-auth-app-django==2.0.0
social-auth-core==1.5.0
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
urllib3==1.22
whitenoise==3.3.1
